pyproject.toml
[tool.poetry]
name = "mcp-server"
version = "0.1.0"
description = "An MCP server providing blog post generation capabilities"
authors = ["Ian Roberts <ian2roberts@gmail.com>"]
readme = "README.md"
license = "MIT"
repository = "https://github.com/ian1roberts/mcp_server"
keywords = ["mcp", "blog", "ai", "openai"]

[tool.poetry.dependencies]
python = "^3.12"
fastmcp = "^2.3.0"
openai = "^1.74.0"
python-dotenv = "^1.0.0"
requests = "^2.28.0"
beautifulsoup4 = "^4.12.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
black = "^23.3.0"
isort = "^5.12.0"
mypy = "^1.2.0"
pytest-cov = "^4.1.0"

[tool.poetry.scripts]
mcp-server = "mcp_server.server:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true